requirements.txt
# Các thư viện cần có để chạy ứng dụng.
streamlit                   # Để chạy web app
mysql-connector-python      # Để kết nối Docker MySQL
pandas                      # Xử lý dữ liệu bảng
matplotlib                  # Vẽ biểu đồ tĩnh
plotly                      # Vẽ biểu đồ tương tác
python-dotenv               # Để quản lý biến môi trường
reportlab
python-pptx
